feat: extract card assignment logic out of `WorkspaceCompanyCardsPage`
